feat(interconnect): support 256B read transactions
Expand upstream read transactions to 256B while keeping the write path at 32B.

Update AXI4 and AXI3 read handling plus unit tests for 64B and 256B requests.

Tested with ctest --test-dir build_axi_kit256 --output-on-failure (6/6 passed).

+++ axi_interconnect/AXI_Interconnect.cpp
@@ -532,9 +532,9 @@ void AXI_Interconnect::debug_print() {
 // Helpers
 // ============================================================================
 uint8_t AXI_Interconnect::calc_burst_len(uint8_t total_size) {
-  uint8_t bytes = total_size + 1;
-  uint8_t beats = (bytes + 3) / 4;
-  return beats > 0 ? beats - 1 : 0;
+  uint16_t bytes = static_cast<uint16_t>(total_size) + 1u;
+  uint16_t beats = (bytes + 3u) / 4u;
+  return beats > 0 ? static_cast<uint8_t>(beats - 1u) : 0;
 }
 
 } // namespace axi_interconnect
